Moved some DOFManager related stuff in DiscretizationUtils header
